1 - Criado o objeto vector transitionList, este será a nossa lista de transições No fonte do fade e do MFIT eu já fiz um exemplo de como manipular este objeto. Para mais informações: http://www.cplusplus.com/reference/stl/vector/
2 - Fiz algumas alterações no 'n' , agora ele imprime o numero do frame corrente na própria janela :D

+++ trunk/DEV/source/Histogram.cpp
@@ -58,7 +58,7 @@ Histogram::Histogram(IplImage* img)
 
 	float max = 0, min = 0;
 	int maxLum = 0, minLum = 0;
-	int maxLuminance = 0;
+	double maxLuminance = 0;
 
 	double soma = 0;
